bugfix that sometimes caused k=0 advance to not work

+++ Firmware/gerry09_iLQR_t4/src/state_estimators/state_estimator_kf.h
@@ -76,7 +76,7 @@ float StateEstimatorKf::stateUpdate(float traj_time_s) {
     state_.k = -1;
   }
 
-  if ((k != state_.k) && (k != (state_.k + 1))) {
+  if ((k != state_.k) && (k != static_cast<int64_t>(state_.k + 1))) {
     // ERROR
     SerialD.printf("Estimator Error:  k = %d, state_.k = %d\n", k, state_.k);
     std::fill(std::begin(state_.xHat), std::end(state_.xHat), 0);
